requirements.txt
gql[requests]==3.4.0
prance[osv]==0.21.2
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.17.1
rich==11.2.0
requests==2.28.1
